Actividad 4 taller 7 de septiembre y controlador de reserva

+++ demo/src/main/java/com/example/demo/repository/EspacioRepository.java
@@ -1,11 +1,9 @@
 package com.example.demo.repository;
 
+import com.example.demo.entities.Espacio;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.demo.entities.Espacio;
-
 @Repository
-public interface EspacioRepository extends JpaRepository<Espacio, Integer> {
-    // JpaRepository ya trae las operaciones básicas para Espacio.
+public interface EspacioRepository extends JpaRepository<Espacio, Long> {
 }
