+++ demo/src/main/java/com/example/demo/repository/ReservaRepository.java
@@ -1,11 +1,9 @@
 package com.example.demo.repository;
 
+import com.example.demo.entities.Reserva;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.demo.entities.Reserva;
-
 @Repository
-public interface ReservaRepository extends JpaRepository<Reserva, Integer> {
-    // Centraliza el acceso a la tabla de reservas.
+public interface ReservaRepository extends JpaRepository<Reserva, Long> {
 }
